Add: 상수데이터 및 Mock데이터 구현
