+++ src/pages/Suhyeon/Main/components/CommentBox.scss
@@ -5,7 +5,7 @@ $context-size: 15px;
     display: flex;
     align-items: center;
     padding-left: 20px;
-    padding-bottom: 10px;
+    padding-top: 10px;
     font-size: $context-size;
 
     .commentId {
