+++ src/pages/Suhyeon/Main/components/FeedContext.scss
@@ -1,0 +1,122 @@
+$context-size: 15px;
+
+.authorProfileBox {
+  display: flex;
+  align-items: center;
+  height: 30px;
+  padding: 40px 20px;
+  border-top: 1px solid lightgray;
+
+  .circle {
+    display: flex;
+    width: 50px;
+    height: 50px;
+    border-radius: 70%;
+    overflow: hidden;
+    justify-content: center;
+    align-items: center;
+
+    .authorProfile_img {
+      width: 70px;
+    }
+  }
+
+  .authorId {
+    font-weight: 800;
+    margin-left: 15px;
+  }
+
+  .otherIcon {
+    width: 20px;
+    margin-left: auto;
+    opacity: 50%;
+  }
+}
+
+.articleImgBox {
+  width: 100%;
+  height: 500px;
+
+  .article_img {
+    width: 100%;
+    height: 100%;
+  }
+}
+
+.reactionIconBox {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  height: 50px;
+  padding: 30px 20px;
+
+  .reactionIcons {
+    width: 30px;
+    margin-right: 5px;
+    opacity: 60%;
+  }
+
+  .reaction_bookmarkIcon {
+    width: 30px;
+    opacity: 30%;
+  }
+}
+
+.statusBox {
+  display: flex;
+  align-items: center;
+  margin-left: 20px;
+  height: 40px;
+
+  .circle_2 {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    align-content: center;
+    width: 30px;
+    height: 30px;
+    border-radius: 70%;
+    overflow: hidden;
+    border-width: 10px;
+
+    .statusProfile_img {
+      width: 100%;
+    }
+  }
+
+  .status {
+    margin-left: 10px;
+    font-weight: 700;
+  }
+}
+
+.postBox {
+  display: flex;
+  justify-items: center;
+  align-content: center;
+  padding-left: 20px;
+  padding-top: 10px;
+  padding-bottom: 10px;
+  font-size: $context-size;
+
+  .postId {
+    font-weight: 800;
+  }
+
+  .postContext {
+    width: 300px;
+    margin-left: 10px;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+  }
+
+  #postMore {
+    margin-left: 0px;
+    border-width: 0px;
+    height: 15px;
+    color: lightgray;
+    font-size: $context-size;
+    background-color: transparent;
+  }
+}
